requirements.txt
# Vuln fixes
pip == 23.3
setuptools == 70.0.0

# Requirements
discord.py == 2.4.0
openpyxl == 3.1.5